requirements.txt
astroid==1.5.3
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.6.16
chardet==3.0.4
colorama==0.3.9
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
Django==1.11.7
djangorestframework==3.10.3
entrypoints==0.3
esridump==1.8.0
geojson==2.4.1
googlemaps==3.0.2
idna==2.8
inflect==2.1.0
ipykernel==5.1.1
ipython==7.7.0
ipython-genutils==0.2.0
ipywidgets==7.5.0
isort==4.2.15
jedi==0.14.1
Jinja2==2.10.1
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
lazy-object-proxy==1.3.1
MapProxy==1.12.0
MarkupSafe==1.1.1
matplotlib==2.1.0
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.5.0
nbformat==4.4.0
notebook==6.0.0
numpy==1.13.3
pandas==0.24.2
pandocfilters==1.4.2
parso==0.5.1
pickleshare==0.7.5
Pillow==6.2.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
psycopg2==2.7.3.2
Pygments==2.4.2
pylint==1.7.4
pyparsing==2.2.0
pyproj==2.4.0
pyrsistent==0.15.4
python-dateutil==2.6.1
pytz==2017.2
pywinpty==0.5.5
PyYAML==5.1.2
pyzmq==18.0.2
qtconsole==4.5.2
requests==2.22.0
schedule==0.5.0
Send2Trash==1.5.0
simplejson==3.16.0
six==1.11.0
sqlacodegen==2.1.0
SQLAlchemy==1.3.6
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.2
urllib3==1.25.3
virtualenv==16.7.2
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.5.0
wrapt==1.10.11
